// src/types/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type MakeEmpty<T extends { [key: string]: unknown }, K extends keyof T> = { [_ in K]?: never };
export type Incremental<T> = T | { [P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
};

export type FileCacheKeys = {
	__typename?: 'FileCacheKeys';
	host?: Maybe<Scalars['String']['output']>;
	key: Scalars['String']['output'];
};

export enum FileType {
	Css = 'CSS',
	Js = 'JS',
}

export type Files = {
	__typename?: 'Files';
	fileName?: Maybe<Scalars['String']['output']>;
	type: FileType;
	urlPrefix: Scalars['String']['output'];
};

export type ProductImages = {
	__typename?: 'ProductImages';
	hints: Array<Maybe<Scalars['String']['output']>>;
	pageUrl?: Maybe<Scalars['String']['output']>;
};